GET Profile Info -> Console
Added a GET for the Player's profile information.  Now it's time to decide what all I want the app to do and what components I need to request.  If I'm not careful I will end up with my own version of callback hell, but instead of callbacks it will be a runtime based exclusively on setters.  I should probably migrate to Promises, but I'm not accustomed to that level of asynchronous JavaScript.  I'll probably decide what to do about it in the next few commits.  Worst case scenario I end up with a moderate refactor to speed up execution time.